2mapping edit

+++ ruoyi-system/src/main/java/com/ruoyi/system/service/impl/SysOssServiceImpl.java
@@ -1,13 +1,17 @@
 package com.ruoyi.system.service.impl;
 
+import cn.dev33.satoken.stp.StpUtil;
 import cn.hutool.core.convert.Convert;
+import cn.hutool.core.io.FileUtil;
 import cn.hutool.core.io.IoUtil;
 import cn.hutool.core.util.ObjectUtil;
+import cn.hutool.http.HttpUtil;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.ruoyi.common.constant.CacheNames;
 import com.ruoyi.common.core.domain.PageQuery;
+import com.ruoyi.common.core.domain.entity.SysUser;
 import com.ruoyi.common.core.page.TableDataInfo;
 import com.ruoyi.common.core.service.OssService;
 import com.ruoyi.common.exception.ServiceException;
@@ -21,8 +25,10 @@ import com.ruoyi.oss.enumd.AccessPolicyType;
 import com.ruoyi.oss.factory.OssFactory;
 import com.ruoyi.system.domain.SysOss;
 import com.ruoyi.system.domain.bo.SysOssBo;
+import com.ruoyi.system.domain.entity.OssFile;
 import com.ruoyi.system.domain.vo.SysOssVo;
 import com.ruoyi.system.mapper.SysOssMapper;
+import com.ruoyi.system.mapper.SysUserMapper;
 import com.ruoyi.system.service.ISysOssService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.cache.annotation.Cacheable;
@@ -31,8 +37,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.io.InputStream;
+import java.io.*;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -46,6 +51,7 @@ import java.util.stream.Collectors;
 public class SysOssServiceImpl implements ISysOssService, OssService {
 
     private final SysOssMapper baseMapper;
+    private final SysUserMapper userMapper;
 
     @Override
     public TableDataInfo<SysOssVo> queryPageList(SysOssBo bo, PageQuery pageQuery) {
@@ -118,6 +124,79 @@ public class SysOssServiceImpl implements ISysOssService, OssService {
         }
     }
 
+
+    /**
+     * @param ossIds
+     * @throws IOException
+     */
+    @Override
+    public List<OssFile> downloadBatchFiles(Collection<Long> ossIds) throws IOException {
+        System.out.println(ossIds.toString());
+        List<SysOssVo> SysOssVoList = SpringUtils.getAopProxy(this).listByIds(ossIds);
+        if (SysOssVoList.isEmpty()) {
+            throw new ServiceException("文件数据不存在!");
+        }
+
+        OssClient storage = OssFactory.instance();
+        List<OssFile> ossFiles = new ArrayList<>();
+        for (SysOssVo sysOssVo:SysOssVoList) {
+            System.out.println(sysOssVo.getUrl());
+            InputStream storageObjectContent = null;
+            for (int retryTimes = 0; retryTimes <= 5 && storageObjectContent == null; retryTimes++){
+                try{
+                    Thread.sleep(200);
+                    storageObjectContent = storage.getObjectContent(sysOssVo.getUrl());
+                }catch (Exception e){
+                    System.out.println("retryTimes is " + retryTimes);
+                }
+            }
+            OssFile ossFile = new OssFile();
+            ossFile.setFileContent(storageObjectContent);
+            ossFile.setSysOssVo(sysOssVo);
+            ossFiles.add(ossFile);
+        }
+
+        return ossFiles;
+    }
+
+
+    /**
+     * @param ossIds
+     * @throws IOException
+     */
+    @Override
+    public List<OssFile> downloadBatchFilesByHttp(Collection<Long> ossIds) throws IOException {
+        System.out.println(ossIds.toString());
+        List<SysOssVo> SysOssVoList = SpringUtils.getAopProxy(this).listByIds(ossIds);
+        if (SysOssVoList.isEmpty()) {
+            throw new ServiceException("文件数据不存在!");
+        }
+
+
+        List<OssFile> ossFiles = new ArrayList<>();
+        for (SysOssVo sysOssVo:SysOssVoList) {
+            System.out.println(sysOssVo.getUrl());
+            ByteArrayOutputStream os = new ByteArrayOutputStream();
+            ByteArrayInputStream storageObjectContent = new ByteArrayInputStream(new byte[10]);
+
+            try{
+                String url = sysOssVo.getUrl().replace("http://kjjs-443.webvpn.zust.edu.cn/minio-oss/", "http://127.0.0.1:9000/");
+                System.out.println(url);
+                HttpUtil.download(url, os, false);
+                storageObjectContent = new ByteArrayInputStream(os.toByteArray());
+            }catch (Exception ignored){
+
+            }
+            OssFile ossFile = new OssFile();
+            ossFile.setFileContent(storageObjectContent);
+            ossFile.setSysOssVo(sysOssVo);
+            ossFiles.add(ossFile);
+        }
+
+        return ossFiles;
+    }
+
+
     @Override
     public SysOssVo upload(MultipartFile file) {
         String originalfileName = file.getOriginalFilename();
@@ -142,17 +221,62 @@ public class SysOssServiceImpl implements ISysOssService, OssService {
         return this.matchingUrl(sysOssVo);
     }
 
+
+    @Override
+    public SysOssVo upload(File file, String originalfileName) {
+        byte[] bytes = new byte[(int) file.length()];
+        try {
+            FileInputStream fis = new FileInputStream(file);
+            fis.read(bytes);
+        } catch (FileNotFoundException e) {
+            throw new ServiceException(e.getMessage());
+        } catch (IOException e) {
+            throw new ServiceException(e.getMessage());
+        }
+        String suffix = StringUtils.substring(originalfileName, originalfileName.lastIndexOf("."), originalfileName.length());
+        OssClient storage = OssFactory.instance();
+        UploadResult uploadResult;
+        uploadResult = storage.uploadSuffix(bytes, suffix, null);
+        // 保存文件信息
+        SysOss oss = new SysOss();
+        oss.setUrl(uploadResult.getUrl());
+        oss.setFileSuffix(suffix);
+        oss.setFileName(uploadResult.getFilename());
+        oss.setOriginalName(originalfileName);
+        oss.setService(storage.getConfigKey());
+        baseMapper.insert(oss);
+        SysOssVo sysOssVo = new SysOssVo();
+        BeanCopyUtils.copy(oss, sysOssVo);
+        return this.matchingUrl(sysOssVo);
+    }
+
+
     @Override
     public Boolean deleteWithValidByIds(Collection<Long> ids, Boolean isValid) {
+        List<Long> idList = new ArrayList<>();
         if (isValid) {
-            // 做一些业务上的校验,判断是否需要校验
+            // 只允许上传者删除oss文件，其它用户只能查看不能删除，用户删除非自己上传的文件不会报错、不会影响其它自己上传的文件的删除
+            LambdaQueryWrapper<SysOss> lqw1 = new LambdaQueryWrapper<>();
+            lqw1.in(SysOss::getOssId,ids);
+            List<SysOss> ossList = baseMapper.selectList(lqw1);
+            SysUser user = userMapper.selectById(StpUtil.getLoginIdAsString().substring(9));
+            ossList.forEach(oss -> {
+                System.out.println(oss.toString());
+                if(oss.getCreateBy().equals(user.getUserName())){
+                    idList.add(oss.getOssId());
+                }
+            });
         }
-        List<SysOss> list = baseMapper.selectBatchIds(ids);
-        for (SysOss sysOss : list) {
-            OssClient storage = OssFactory.instance(sysOss.getService());
-            storage.delete(sysOss.getUrl());
+        if (!idList.isEmpty()){
+            List<SysOss> list = baseMapper.selectBatchIds(idList);
+            for (SysOss sysOss : list) {
+                OssClient storage = OssFactory.instance(sysOss.getService());
+                storage.delete(sysOss.getUrl());
+            }
+            return baseMapper.deleteBatchIds(idList) > 0;
+        }else{
+            return false;
         }
-        return baseMapper.deleteBatchIds(ids) > 0;
     }
 
     /**
@@ -162,11 +286,13 @@ public class SysOssServiceImpl implements ISysOssService, OssService {
      * @return oss 匹配Url的OSS对象
      */
     private SysOssVo matchingUrl(SysOssVo oss) {
-        OssClient storage = OssFactory.instance(oss.getService());
-        // 仅修改桶类型为 private 的URL，临时URL时长为120s
-        if (AccessPolicyType.PRIVATE == storage.getAccessPolicy()) {
-            oss.setUrl(storage.getPrivateUrl(oss.getFileName(), 120));
-        }
+        // 没有用到临时url，这里注释掉。
+//      // 否则每次调用有关文件的方法时都会实例化工厂对象去匹配。
+//        OssClient storage = OssFactory.instance(oss.getService());
+//        // 仅修改桶类型为 private 的URL，临时URL时长为120s
+//        if (AccessPolicyType.PRIVATE == storage.getAccessPolicy()) {
+//            oss.setUrl(storage.getPrivateUrl(oss.getFileName(), 120));
+//        }
         return oss;
     }
 }
